rangefeed: move rangefeed test helpers into separate file
This PR moves some rangefeed test helpers into a separate file so some
can be exported.

Epic: None

Release notes: None
