+++ pkg/kv/kvserver/rangefeed/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "stream_muxer.go",
         "stream_muxer_test_helper.go",
         "task.go",
+        "test_helpers.go",
         "testutil.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/rangefeed",
@@ -30,6 +31,8 @@ go_library(
         "//pkg/roachpb",
         "//pkg/settings",
         "//pkg/settings/cluster",
+        "//pkg/sql/rowenc/keyside",
+        "//pkg/sql/sem/tree",
         "//pkg/storage",
         "//pkg/storage/enginepb",
         "//pkg/storage/fs",
@@ -37,6 +40,7 @@ go_library(
         "//pkg/util/bufalloc",
         "//pkg/util/buildutil",
         "//pkg/util/container/heap",
+        "//pkg/util/encoding",
         "//pkg/util/envutil",
         "//pkg/util/hlc",
         "//pkg/util/interval",
@@ -80,7 +84,6 @@ go_test(
         "//pkg/settings/cluster",
         "//pkg/sql/randgen",
         "//pkg/sql/rowenc/keyside",
-        "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/storage",
         "//pkg/storage/enginepb",
